+++ Project3/GameState_Playground.h
@@ -33,7 +33,7 @@ private:
 
 	int score = 0;
 	bool setted = false;
-	double timeInGame = 0.f;
+	float timeInGame = 0.f;
 
 
 	//// entities
@@ -64,9 +64,9 @@ private:
 	bool erased = false;
 
 	// x coordinate for success pole
-	double successPoleXCoordinate;
+	float successPoleXCoordinate;
 
-	double successTime = 0.f;
+	float successTime = 0.f;
 
 	sf::Texture texture;
 	sf::View view;
