+++ Project3/NyanCat.h
@@ -8,9 +8,9 @@ private:
 
 	// status
 	//time since the last time the entity took damage
-	double timeSinceHurt = 0.f;
+	float timeSinceHurt = 0.f;
 	//minimum time between taking damage
-	double minHurtTimeGap = 1.5f;
+	float minHurtTimeGap = 1.5f;
 
 
 	//initializer functions
@@ -18,8 +18,8 @@ private:
 	void initComponents();
 
 	// implement flicker
-	const double flickerTimeGap = 0.1f;
-	double currentFlickerTime = 0.f;
+	const float flickerTimeGap = 0.1f;
+	float currentFlickerTime = 0.f;
 	bool flickerNoRender;
 
 	float currenDt = 0.f;
